[4493] (int -> double) by nik.
git-svn-id: http://svn.l2jserver.com/trunk@4497 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2J_Server/java/com/l2jserver/gameserver/skills/hitConditionBonus.java
@@ -46,7 +46,7 @@ public class hitConditionBonus
 	
 	protected static double getConditionBonus(L2Character attacker, L2Character target)
 	{
-		int mod = 100;
+		double mod = 100;
 		// Get high or low bonus
 		if (attacker.getZ() - target.getZ() > 50)
 			mod += hitConditionBonus.highBonus;
